refactor: rename path component
