+++ src/components/Pagination/style.ts
