+++ src/components/RemedyView/components/DocumentTable/index.tsx
