+++ src/components/RemedyView/index.tsx
