+++ src/components/TableListRemedies/index.tsx
