+++ src/components/TableListRemedies/styles.ts
