+++ src/layouts/DefaultLayout.tsx
@@ -1,5 +1,5 @@
 import { Toaster } from 'sonner'
-import { Header } from '../Components/Header'
+import { Header } from '../components/Header'
 import { Outlet } from 'react-router-dom'
 
 export function DefaultLayout() {
