+++ src/pages/NewRemedy/index.tsx
@@ -1,17 +1,17 @@
 import { useForm } from 'react-hook-form'
-import { Button } from '../../Components/Button'
-import { FormDocument } from '../../Components/FormDocument'
-import ModalDialog from '../../Components/ModalDialog'
+import { Button } from '../../components/Button'
+import { FormDocument } from '../../components/FormDocument'
+import ModalDialog from '../../components/ModalDialog'
 import { HeaderDocuments, NewRemedyContainer } from './styles'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { FormPrincipleAtive } from '../../Components/FormPricipleActive'
-import ListDocuments from '../../Components/ListDocuments'
-import ListPrincipleActives from '../../Components/ListPrincipleActives'
+import { FormPrincipleAtive } from '../../components/FormPricipleActive'
+import ListDocuments from '../../components/ListDocuments'
+import ListPrincipleActives from '../../components/ListPrincipleActives'
 import { useRemedies } from '../../hooks/useRemedies'
 import { toast } from 'sonner'
 import { Link } from 'react-router-dom'
-import { FormRemedy } from '../../Components/FormRemedy'
+import { FormRemedy } from '../../components/FormRemedy'
 import { useEffect, useState } from 'react'
 
 const formRemedySchema = z.object({
