+++ src/pages/UpdateRemedy/index.tsx
@@ -1,14 +1,14 @@
-import { Button } from '../../Components/Button'
-import { FormDocument } from '../../Components/FormDocument'
-import ModalDialog from '../../Components/ModalDialog'
+import { Button } from '../../components/Button'
+import { FormDocument } from '../../components/FormDocument'
+import ModalDialog from '../../components/ModalDialog'
 import { HeaderDocuments, UpdateFormContainer } from './styles'
-import { FormPrincipleAtive } from '../../Components/FormPricipleActive'
-import ListDocuments from '../../Components/ListDocuments'
-import ListPrincipleActives from '../../Components/ListPrincipleActives'
+import { FormPrincipleAtive } from '../../components/FormPricipleActive'
+import ListDocuments from '../../components/ListDocuments'
+import ListPrincipleActives from '../../components/ListPrincipleActives'
 import { useCallback, useEffect } from 'react'
 import { useRemedies } from '../../hooks/useRemedies'
 import { Link, useParams } from 'react-router-dom'
-import { FormRemedy } from '../../Components/FormRemedy'
+import { FormRemedy } from '../../components/FormRemedy'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useForm } from 'react-hook-form'
 import { z } from 'zod'
